Move storageClass functionallity from tests/utils to libstorage
